Use template repository to fetch template for email
Define a repository that will fetch a template for an email. The key for
the template in the storage would be the class name of the email. This
way, we won't need to have an enum with all email types. And the method
that will fetch an email template will be universal.